# docs/swagger.yaml
openapi: 3.0.0
info:
  title: GameRent API
  version: 1.0.0
  description: >
    REST API specification for the GameRent platform. Supports game account rentals,
    user trust level management, reviews, notifications, admin operations, and payment webhooks.
servers:
  - url: /api/v1
    description: Production API Server (v1)
paths:
...
  /me/rentals:
    get:
      summary: Get current user rentals
      description: Retrieves rentals corresponding to the authenticated user.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Rentals retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRentalsResponseWrapped'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/payments:
    get:
      summary: Get current user payments
      description: Retrieves payments history corresponding to the authenticated user.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Payments retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPaymentsResponseWrapped'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/notifications:
    get:
      summary: Get current user notifications
      description: Retrieves notifications list corresponding to the authenticated user.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Notifications retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotificationsResponseWrapped'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /accounts/{id}/favorite:
    post:
      summary: Add account to favorites
      description: Adds a game account to the user's favorites list.
      tags:
        - Accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Added to favorites.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Remove account from favorites
      description: Removes a game account from the user's favorites list.
      tags:
        - Accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Removed from favorites.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /rentals:
    post:
      summary: Create a rental
      description: Books a game account for renting and calculates costs.
      tags:
        - Rentals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRentalRequest'
      responses:
        '201':
          description: Rental created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalResponseWrapped'
        '400':
          description: Invalid rental duration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden due to insufficient trust level.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Account already rented or unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List rentals
      description: Retrieves a list of rentals, filtered by status.
      tags:
        - Rentals
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [Created, WaitingPayment, Active, Expired, Completed, Cancelled]
      responses:
        '200':
          description: Rentals list retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRentalsResponseWrapped'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rentals/{rentalId}:
    get:
      summary: Get rental by ID
      description: Retrieves details of a specific rental.
      tags:
        - Rentals
      security:
        - bearerAuth: []
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Rental details retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalResponseWrapped'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rental not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rentals/{rentalId}/cancel:
    post:
      summary: Cancel a rental
      description: Cancels a rental in the "WaitingPayment" or "Created" state.
      tags:
        - Rentals
      security:
        - bearerAuth: []
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Rental cancelled successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalResponseWrapped'
        '400':
          description: Cannot cancel rental in current state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rentals/calculate:
    post:
      summary: Calculate rental cost
      description: Estimates hourly, deposit, and total costs for a hypothetical rental duration.
      tags:
        - Rentals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateRentalRequest'
      responses:
        '200':
          description: Rental cost calculation details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateRentalResponseWrapped'
        '400':
          description: Invalid request parameter values.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /payments/{paymentId}:
    get:
      summary: Get payment by ID
      description: Retrieves details of a specific payment transaction.
      tags:
        - Payments
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment details retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseWrapped'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Payment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          example: {}
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: The input parameters are invalid.

    # Wrapped DTO Response schemas
...
    PaginatedRentalsResponseWrapped:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PaginatedRentalsResponse'
...
    PaginatedPaymentsResponseWrapped:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PaginatedPaymentsResponse'
...
    PaginatedNotificationsResponseWrapped:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PaginatedNotificationsResponse'
...